chore(tests): add 'record' npm script to functional-tests
