+++ packages/functional-tests/tests/stub.spec.ts
@@ -1,0 +1,6 @@
+import { test } from '../lib/fixtures/standard';
+
+test('stub', async ({ credentials, page }) => {
+  console.log(credentials);
+  await page.pause();
+});
